+++ app/src/main/java/com/example/projet_progm/ActivitySearchTheChest.kt
@@ -48,7 +48,7 @@ class ActivitySearchTheChest : ComponentActivity() {
     private val endGame = object : Runnable {
         override fun run() {
             val resultIntent = Intent()
-            resultIntent.putExtra("score", score)
+            resultIntent.putExtra("score", score.toInt())
             setResult(Activity.RESULT_OK, resultIntent)
             finish()
         }
