+++ app/src/main/java/com/example/projet_progm/EnigmeActivity.kt
@@ -71,7 +71,7 @@ class EnigmeActivity : ComponentActivity() {
     private val endGame = object : Runnable {
         override fun run() {
             val resultIntent = Intent()
-            resultIntent.putExtra("score", score)
+            resultIntent.putExtra("score", score.toInt())
             setResult(Activity.RESULT_OK, resultIntent)
             finish()
         }
